Update frontend to use PageRouter for layout structure
Now we have top, content and bottom of the page more clearly divided into components.
