+++ frontend/src/components/uiComponents/AppContent.js
@@ -10,8 +10,8 @@
  * @requires components/RetentionRate
  */
 import React from 'react'
+import { BrowserRouter as Route, Switch } from 'react-router-dom'
 
-import LoginForm from '../LoginForm'
 import Users from '../Users'
 import Caregivers from '../Caregivers'
 import Cumulative from '../Cumulative'
@@ -23,15 +23,7 @@ import RetentionRate from '../RetentionRate'
  * @param {*} param0 - all props from App.js
  * @returns {object} - JSX component containing all sub components
  */
-const AppContent = ({ user, appUsers, caregivers, cumulativeUsers, activeUsers, retentionRates, averageRetention,
-  username, setUsername, password, setPassword, setUser }) => {
-
-  const containerStyle = {
-    position: 'relative',
-    minHeight: '100vh',
-    backgroundColor: '#f2f2f2',
-    paddingBottom: '100px'
-  }
+const AppContent = ({ appUsers, caregivers, cumulativeUsers, activeUsers, retentionRates, averageRetention }) => {
 
   const subContainer1 = {
     marginTop: '10px',
@@ -45,19 +37,33 @@ const AppContent = ({ user, appUsers, caregivers, cumulativeUsers, activeUsers, 
     justifyContent: 'center',
     alignItems: 'center'
   }
-  return (
-    <div style={containerStyle}>
-      {user
-        ? null
-        : <React.Fragment>
-          <LoginForm username={username} setUsername={setUsername} password={password}
-            setPassword={setPassword} user={user} setUser={setUser} />
-        </React.Fragment>
-      }
 
-      {user
-        ?
-        <React.Fragment>
+  const centered = {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center'
+  }
+
+  return (
+    <div className="body">
+      <Switch>
+
+        <Route path='/retention'>
+          <div style={centered}>
+            <RetentionRate
+              retentionRates={retentionRates}
+              average={averageRetention} />
+          </div>
+        </Route>
+
+        <Route path='/cumulative'>
+          <div style={centered}>
+            <Cumulative cumulative={cumulativeUsers}
+              activeUsers={activeUsers} />
+          </div>
+        </Route>
+
+        <Route path='/'>
           <div className="p-grid p-fluid dashboard" style={subContainer1}>
             <Users users={appUsers} />
             <Caregivers caregivers={caregivers} />
@@ -72,11 +78,9 @@ const AppContent = ({ user, appUsers, caregivers, cumulativeUsers, activeUsers, 
                 average={averageRetention} />
             </div>
           </div>
-        </React.Fragment>
-        :
-        null}
-      <div>
-      </div>
+        </Route>
+
+      </Switch>
     </div>
   )
 }
