+++ frontend/src/components/uiComponents/AppTopbar.js
@@ -17,12 +17,7 @@ import { Toolbar } from 'primereact/toolbar'
 import { Button } from 'primereact/button'
 import loginService from '../../services/loginService'
 
-import { Route } from 'react-router-dom'
-
 import Header from './AppHeader'
-import AppContent from './AppContent'
-import RetentionRate from '../RetentionRate'
-import Cumulative from '../Cumulative'
 
 /**
  * Component for applications UI topbar
@@ -30,8 +25,7 @@ import Cumulative from '../Cumulative'
  * @param {*} param0 - all props from App.js
  * @returns {object} - JSX Topbar component
  */
-const AppTopbar = ({ user, appUsers, caregiverFilterForAllUsers, handleFilterChange, caregivers, cumulativeUsers, activeUsers, retentionRates, averageRetention,
-  username, setUsername, password, setPassword, setUser }) => {
+const AppTopbar = ({ user, setUser, caregiverFilterForAllUsers, handleFilterChange }) => {
 
   /**
    * Handle logout button presses
@@ -96,44 +90,8 @@ const AppTopbar = ({ user, appUsers, caregiverFilterForAllUsers, handleFilterCha
     </React.Fragment>
   )
 
-  const centered = {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center'
-  }
-
   return (
-    <div className="p-component">
-      <Toolbar left={leftContents} right={rightContents} style={toolbarStyle} />
-      <Route path='/retention'>
-        <div style={centered}>
-          <RetentionRate retentionRates={retentionRates}
-            averageRetention={averageRetention} />
-        </div>
-      </Route>
-      <Route path='/cumulative'>
-        <div style={centered}>
-          <Cumulative cumulative={cumulativeUsers}
-            activeUsers={activeUsers} />
-        </div>
-      </Route>
-      <Route path='/'>
-        <AppContent user={user}
-          appUsers={appUsers}
-          caregivers={caregivers}
-          caregiverFilterForAllUsers={caregiverFilterForAllUsers}
-          handleFilterChange={handleFilterChange}
-          cumulativeUsers={cumulativeUsers}
-          activeUsers={activeUsers}
-          retentionRates={retentionRates}
-          averageRetention={averageRetention}
-          username={username}
-          setUsername={setUsername}
-          password={password}
-          setPassword={setPassword}
-          setUser={setUser} />
-      </Route>
-    </div>
+    <Toolbar left={leftContents} right={rightContents} style={toolbarStyle} />
   )
 }
